requirements.txt
pyftpdlib==1.5.7
pycryptodome==3.17